atlar: create client package, remove metrics and add others
