+++ components/payments/cmd/connectors/internal/connectors/atlar/client/accounts.go
@@ -1,0 +1,23 @@
+package client
+
+import (
+	"context"
+	"time"
+
+	"github.com/formancehq/payments/cmd/connectors/internal/connectors"
+	"github.com/get-momo/atlar-v1-go-client/client/accounts"
+)
+
+func (c *Client) GetV1Accounts(ctx context.Context, token string, pageSize int64) (*accounts.GetV1AccountsOK, error) {
+	f := connectors.ClientMetrics(ctx, "atlar", "list_accounts")
+	now := time.Now()
+	defer f(ctx, now)
+
+	accountsParams := accounts.GetV1AccountsParams{
+		Limit:   &pageSize,
+		Context: ctx,
+		Token:   &token,
+	}
+
+	return c.client.Accounts.GetV1Accounts(&accountsParams)
+}
