+++ components/payments/cmd/connectors/internal/connectors/atlar/client/utils_test.go
@@ -1,4 +1,4 @@
-package atlar
+package client
 
 import (
 	"errors"
