+++ components/payments/cmd/connectors/internal/connectors/atlar/task_fetch_transactions.go
@@ -8,19 +8,18 @@ import (
 	"math/big"
 	"time"
 
+	"github.com/formancehq/payments/cmd/connectors/internal/connectors/atlar/client"
 	"github.com/formancehq/payments/cmd/connectors/internal/connectors/currency"
 	"github.com/formancehq/payments/cmd/connectors/internal/ingestion"
 	"github.com/formancehq/payments/cmd/connectors/internal/task"
 	"github.com/formancehq/payments/internal/models"
 	"github.com/formancehq/stack/libs/go-libs/contextutil"
 	"github.com/formancehq/stack/libs/go-libs/logging"
-	"github.com/formancehq/stack/libs/go-libs/pointer"
-	atlar_client "github.com/get-momo/atlar-v1-go-client/client"
 	"github.com/get-momo/atlar-v1-go-client/client/transactions"
 	atlar_models "github.com/get-momo/atlar-v1-go-client/models"
 )
 
-func FetchTransactionsTask(config Config, client *atlar_client.Rest) task.Task {
+func FetchTransactionsTask(config Config, client *client.Client) task.Task {
 	return func(
 		ctx context.Context,
 		logger logging.Logger,
@@ -30,17 +29,10 @@ func FetchTransactionsTask(config Config, client *atlar_client.Rest) task.Task {
 		ingester ingestion.Ingester,
 	) error {
 		// Pagination works by cursor token.
-		params := transactions.GetV1TransactionsParams{
-			Limit: pointer.For(int64(config.ApiConfig.PageSize)),
-		}
 		for token := ""; ; {
 			requestCtx, cancel := contextutil.DetachedWithTimeout(ctx, 30*time.Second)
 			defer cancel()
-			params.Context = requestCtx
-			params.Token = &token
-			limit := int64(config.PageSize)
-			params.Limit = &limit
-			pagedTransactions, err := client.Transactions.GetV1Transactions(&params)
+			pagedTransactions, err := client.GetV1Transactions(requestCtx, token, int64(config.PageSize))
 			if err != nil {
 				return err
 			}
